+++ systems/trajectory_optimization/BUILD.bazel
@@ -19,6 +19,24 @@ drake_cc_library(
 )
 
 drake_cc_library(
+    name = "position_constraint_force_evaluator",
+    srcs = ["position_constraint_force_evaluator.cc"],
+    hdrs = ["position_constraint_force_evaluator.h"],
+    deps = [
+        ":generalized_constraint_force_evaluator",
+    ],
+)
+
+drake_cc_library(
+    name = "joint_limit_constraint_force_evaluator",
+    srcs = ["joint_limit_constraint_force_evaluator.cc"],
+    hdrs = ["joint_limit_constraint_force_evaluator.h"],
+    deps = [
+        ":generalized_constraint_force_evaluator",
+    ],
+)
+
+drake_cc_library(
     name = "multiple_shooting",
     srcs = ["multiple_shooting.cc"],
     hdrs = ["multiple_shooting.h"],
@@ -98,17 +116,39 @@ drake_cc_googletest(
     ],
 )
 
-drake_cc_googletest(
-    name = "generalized_constraint_force_evaluator_test",
+drake_cc_library(
+    name = "generalized_constraint_force_evaluator_test_util",
+    testonly = 1,
+    srcs = [
+        "test/generalized_constraint_force_evaluator_test_util.cc",
+    ],
+    hdrs = ["test/generalized_constraint_force_evaluator_test_util.h"],
     data = [
         "//examples/simple_four_bar:models",
     ],
     deps = [
-        ":generalized_constraint_force_evaluator",
         "//common:find_resource",
+        "//multibody/parsers",
+    ],
+)
+
+drake_cc_googletest(
+    name = "position_constraint_force_evaluator_test",
+    deps = [
+        ":generalized_constraint_force_evaluator_test_util",
+        ":position_constraint_force_evaluator",
         "//common/test_utilities:eigen_matrix_compare",
         "//math:autodiff",
-        "//multibody/parsers",
+    ],
+)
+
+drake_cc_googletest(
+    name = "joint_limit_constraint_force_evaluator_test",
+    deps = [
+        ":generalized_constraint_force_evaluator_test_util",
+        ":joint_limit_constraint_force_evaluator",
+        "//common/test_utilities:eigen_matrix_compare",
+        "//math:autodiff",
     ],
 )
