+++ systems/trajectory_optimization/test/generalized_constraint_force_evaluator_test_util.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <memory>
+
+#include "drake/multibody/rigid_body_tree.h"
+
+namespace drake {
+namespace systems {
+namespace trajectory_optimization {
+/** Construct a RigidBodyTree containing a four bar linkage. */
+std::unique_ptr<RigidBodyTree<double>> ConstructFourBarTree();
+}  // namespace trajectory_optimization
+}  // namespace systems
+}  // namespace drake
